# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es la API de RPL 2.0 para todos y todas",
        "version": "0.1",
        "title": "RPL 2.0",
        "termsOfService": "Terms of service",
        "contact": {
            "name": "Alejandro Levinas o Mat\u00edas Cano",
            "url": "http://example.com/support",
            "email": "levinasale@gmail.com;matiasjosecc@gmail.com"
        },
        "license": {
            "name": "License of API",
            "url": "http://example.com/support"
        }
    },
  "host": "localhost:38497",
    "basePath": "/",
...
    "paths": {
...
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "authentication-controller"
                ],
                "summary": "getUser",
                "operationId": "getUserUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "accountNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "accountNonLocked",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "credentialsNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "enabled",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.dateCreated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.degree",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.emailValidated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "user.id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "user.lastUpdated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.studentId",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.surname",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.university",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserResponseDTO"
                        }
                    },
                    "403": {
                        "description": "Forbidden!!!!!"
                    },
                    "500": {
                        "description": "500 message",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "deprecated": false
            }
        },
...
        "/api/courses/{courseId}/activities": {
            "get": {
                "tags": [
                    "activities-controller"
                ],
                "summary": "getActivities",
                "operationId": "getActivitiesUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "accountNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "accountNonLocked",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "courseId",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "credentialsNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "enabled",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.dateCreated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.degree",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.emailValidated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "user.id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "user.lastUpdated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.studentId",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.surname",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.university",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserActivityResponseDTO"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden!!!!!"
                    },
                    "500": {
                        "description": "500 message",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": [
                    "activities-controller"
                ],
                "summary": "createActivity",
                "operationId": "createActivityUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "accountNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "accountNonLocked",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "activityCategoryId",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "courseId",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "credentialsNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "enabled",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "initialCode",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "supportingFile",
                        "in": "formData",
                        "description": "supportingFile",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "user.dateCreated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.degree",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.emailValidated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "user.id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "user.lastUpdated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.studentId",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.surname",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.university",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ActivityResponseDTO"
                        }
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "ActivityResponseDTO": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "activity_iotests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/IOTestResponseDTO"
                    }
                },
                "activity_unit_tests": {
                    "type": "string"
                },
                "category_description": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "category_name": {
                    "type": "string"
                },
                "course_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "date_created": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string"
                },
                "file_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "initial_code": {
                    "type": "string"
                },
              "is_iotested": {
                "type": "boolean"
              },
                "language": {
                    "type": "string"
                },
                "last_updated": {
                    "type": "string",
                    "format": "date-time"
                },
                "name": {
                    "type": "string"
                }
            },
            "title": "ActivityResponseDTO"
        },
...
        "IOTestResponseDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "in": {
                    "type": "string"
                },
                "out": {
                    "type": "string"
                }
            },
            "title": "IOTestResponseDTO"
        },
...
        "UserActivityResponseDTO": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "category_description": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "course_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "date_created": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string"
                },
                "file_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
              "is_iotested": {
                "type": "boolean"
              },
                "language": {
                    "type": "string",
                    "enum": [
                        "C",
                        "PYTHON3",
                        "JAVA"
                    ]
                },
                "last_submission_date": {
                    "type": "string",
                    "format": "date-time"
                },
                "last_updated": {
                    "type": "string",
                    "format": "date-time"
                },
                "name": {
                    "type": "string"
                },
                "submission_status": {
                    "type": "string"
                }
            },
            "title": "UserActivityResponseDTO"
        },
...
        "UserResponseDTO": {
            "type": "object",
            "properties": {
                "date_created": {
                    "type": "string",
                    "format": "date-time"
                },
                "degree": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_validated": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "last_updated": {
                    "type": "string",
                    "format": "date-time"
                },
                "name": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            },
            "title": "UserResponseDTO"
        }
    }
}